refactor(terminal): extract TerminalBurstController
- Move wheel-scroll BURST boost (onActiveWheel/onUserScrollIntent), the
  interactive resource-profile override throttle, and write-driven BURST
  tier (onPtyWrite/scheduleWriteBurstDecay) into a new sibling controller
- TerminalInstanceService wires it into makeListenerInstallDeps and the
  write controller's onWrite callback, and delegates its destroy(id)
  cleanup before the instance is removed from the instances map
