+++ src/services/terminal/TerminalInstanceService.ts
@@ -35,6 +35,7 @@ import { TerminalReflowController, forceXtermReflow } from "./TerminalReflowCont
 import { TerminalReconciliationWatchdog } from "./TerminalReconciliationWatchdog";
 import { TerminalWriteController } from "./TerminalWriteController";
 import { TerminalSettleWaiterRegistry } from "./TerminalSettleWaiterRegistry";
+import { TerminalBurstController } from "./TerminalBurstController";
 import { reportFileLinkFailure } from "./FileLinksAddon";
 import {
   installTerminalBoundListeners,
@@ -94,16 +95,6 @@ const WORKER_INGEST_DRAIN_POLL_MS = 10;
 // terminals instead of freezing the renderer for the whole batch.
 const RESIZE_PASS_CHUNK_SIZE = 1;
 
-// Interactive resource-profile override (symptom B): while actively scrolling a
-// full-screen TUI, ask the main process to hold the profile off efficiency.
-// DURATION must exceed THROTTLE so re-requests overlap and the hold never lapses
-// mid-scroll; both are short so the hold self-expires soon after scrolling stops.
-const INTERACTIVE_OVERRIDE_DURATION_MS = 1500;
-const INTERACTIVE_OVERRIDE_THROTTLE_MS = 500;
-// BURST-tier hold after a wheel scroll before reverting to the computed tier —
-// matches the keystroke input-burst decay so a scroll feels just as responsive.
-const WHEEL_BURST_DECAY_MS = 1000;
-
 function canAutoInitializeTerminalIngest(): boolean {
   return (
     typeof window !== "undefined" &&
@@ -121,21 +112,10 @@ class TerminalInstanceService {
   // this after its async addon load and aborts, so a panel torn down mid-build
   // never publishes a stale instance/onData listener for a dead id.
   private cancelledCreations = new Set<string>();
-  // Throttle for the interactive resource-profile override IPC (see onActiveWheel).
-  private lastInteractiveOverrideRequestAt = 0;
-  // Which terminal received the most recent alt-buffer wheel event, and when —
-  // drives the ingest service's background-drain hold during a scroll gesture.
-  // Uses the same decay window as the BURST tier so "gesture over" is one
-  // consistent notion across the renderer.
-  private lastActiveWheelId: string | null = null;
-  private lastActiveWheelAt = 0;
   private dataBuffer = new TerminalOutputIngestService(
     (id, data, chunkCount) => this.writeToTerminal(id, data, chunkCount),
     () => usePanelStore.getState().focusedId,
-    () =>
-      this.lastActiveWheelId !== null && Date.now() - this.lastActiveWheelAt < WHEEL_BURST_DECAY_MS
-        ? this.lastActiveWheelId
-        : null
+    () => this.burstController.getActiveWheelHoldId()
   );
   private suppressedExitUntil = new Map<string, number>();
   private unseenTracker = new TerminalUnseenOutputTracker();
@@ -156,6 +136,7 @@ class TerminalInstanceService {
   private reconciliationWatchdog: TerminalReconciliationWatchdog;
   private writeController: TerminalWriteController;
   private settleWaiters: TerminalSettleWaiterRegistry;
+  private burstController: TerminalBurstController;
   private unsubTierChanged: (() => void) | null = null;
   // Live worker-parse ingest (issue #10960), gated by
   // DAINTREE_PAINT_FABRIC_WORKER_INGEST. One controller per terminal that has
@@ -193,6 +174,11 @@ class TerminalInstanceService {
       writeData: (id, data, chunkCount) => this.writeToTerminal(id, data, chunkCount),
     });
 
+    this.burstController = new TerminalBurstController({
+      getInstance: (id) => this.instances.get(id),
+      applyRendererPolicy: (id, tier) => this.rendererPolicy.applyRendererPolicy(id, tier),
+    });
+
     this.writeController = new TerminalWriteController({
       getInstance: (id) => this.instances.get(id),
       acknowledgePortData: (id, bytes, chunkCount) =>
@@ -201,7 +187,7 @@ class TerminalInstanceService {
       notifyWriteComplete: (id, bytes) => this.dataBuffer.notifyWriteComplete(id, bytes),
       incrementUnseen: (id, isScrolledBack) =>
         this.unseenTracker.incrementUnseen(id, isScrolledBack),
-      onWrite: (id) => this.onPtyWrite(id),
+      onWrite: (id) => this.burstController.onPtyWrite(id),
     });
 
     this.reflowController = new TerminalReflowController({
@@ -379,8 +365,8 @@ class TerminalInstanceService {
       clearDirectingState: (id, trigger) =>
         this.agentStateController.clearDirectingState(id, trigger),
       onUserInput: (id, data) => this.onUserInput(id, data),
-      onActiveWheel: (id) => this.onActiveWheel(id),
-      onUserScrollIntent: (id) => this.onUserScrollIntent(id),
+      onActiveWheel: (id) => this.burstController.onActiveWheel(id),
+      onUserScrollIntent: (id) => this.burstController.onUserScrollIntent(id),
       onEnterPressed: (id) => this.onEnterPressed(id),
       updateLastObservedTitle: (id, title) =>
         usePanelStore.getState().updateLastObservedTitle(id, title),
@@ -616,127 +602,6 @@ class TerminalInstanceService {
     }, 1000);
 
     this.agentStateController.onUserInput(id, data);
-  }
-
-  /**
-   * Active wheel scroll in a focused full-screen mouse-reporting TUI. Scrolling
-   * such a TUI is an app-owned PTY round-trip per line, and a focused-but-idle
-   * pane sits at FOCUSED (10fps): without this the first flick repaints slowly
-   * until the TUI's own redraw output happens to bump the tier. We (1) lift the
-   * renderer to BURST (60fps) for the scroll — reusing the input-burst decay
-   * timer, since a scroll wants the same ~1s revert as a keystroke — and (2) ask
-   * the main process to hold the resource profile at ≥ balanced, so efficiency
-   * mode's stretched port-batch delay (40ms vs 16ms) can't throttle the
-   * returned-redraw stream mid-scroll ("gets slow and stays slow", symptom B).
-   */
-  private onActiveWheel(id: string): void {
-    const managed = this.instances.get(id);
-    if (!managed) return;
-
-    this.lastActiveWheelId = id;
-    this.lastActiveWheelAt = Date.now();
-
-    this.rendererPolicy.applyRendererPolicy(id, TerminalRefreshTier.BURST);
-    if (managed.inputBurstTimer !== undefined) {
-      clearTimeout(managed.inputBurstTimer);
-    }
-    managed.inputBurstTimer = window.setTimeout(() => {
-      const current = this.instances.get(id);
-      if (!current) return;
-      current.inputBurstTimer = undefined;
-      this.rendererPolicy.applyRendererPolicy(id, current.getRefreshTier());
-    }, WHEEL_BURST_DECAY_MS);
-
-    this.requestInteractiveProfileOverride();
-  }
-
-  /**
-   * Ordinary scrollback wheel/key scrolling (not mouse-reporting forwarding).
-   * Holds the resource profile off efficiency for the same reason as
-   * `onActiveWheel` — a profile downgrade mid-scroll drops the WebGL upper
-   * threshold and can force the visible terminal onto the DOM renderer right
-   * as the user is scrolling it (#10858). Unlike `onActiveWheel`, plain
-   * scrollback is client-side xterm rendering with no PTY round-trip per
-   * line, so there's no renderer-tier boost to apply here.
-   */
-  private onUserScrollIntent(id: string): void {
-    const managed = this.instances.get(id);
-    if (!managed) return;
-
-    this.requestInteractiveProfileOverride();
-  }
-
-  /**
-   * Fire-and-forget request that the main process keep the resource profile off
-   * efficiency while the user is actively scrolling. Throttled hard because a
-   * trackpad momentum stream calls this dozens of times/sec; the override window
-   * (INTERACTIVE_OVERRIDE_DURATION_MS) is longer than the throttle so the hold
-   * never lapses between requests, and self-expires shortly after scrolling stops.
-   */
-  private requestInteractiveProfileOverride(): void {
-    const now = Date.now();
-    if (now - this.lastInteractiveOverrideRequestAt < INTERACTIVE_OVERRIDE_THROTTLE_MS) return;
-    this.lastInteractiveOverrideRequestAt = now;
-    try {
-      // safeFireAndForget swallows the ASYNC rejection (channel skew during a
-      // reload, future validation), which a bare synchronous try/catch around a
-      // Promise-returning IPC call would miss; the try/catch still guards a
-      // synchronous throw if `window.electron.system` is absent. Either way the
-      // renderer-side BURST boost above already applied — this is non-critical.
-      safeFireAndForget(
-        window.electron.system.requestInteractiveOverride(INTERACTIVE_OVERRIDE_DURATION_MS),
-        { context: "terminal.requestInteractiveProfileOverride" }
-      );
-    } catch {
-      // window.electron.system unavailable — non-critical.
-    }
-  }
-
-  /**
-   * Write-driven BURST tier: each PTY write extends the burst window in O(1)
-   * by bumping `writeBurstDeadline`. A single self-rearming timer handles
-   * decay — it re-checks the deadline on fire and either reschedules for the
-   * remaining time (if a write extended the window while it was pending) or
-   * reverts the tier via the panel's current `getRefreshTier()`.
-   *
-   * Avoiding per-write clearTimeout/setTimeout matters: at 60fps+ output the
-   * naive pattern thrashes Chromium's timer queue and produces GC pressure.
-   *
-   * `applyRendererPolicy(BURST)` is called on every write: when BURST is
-   * already applied the policy returns early (line 50 of
-   * TerminalRendererPolicy) and as a load-bearing side-effect clears any
-   * pending tierChangeTimer — that cancellation is what prevents a
-   * concurrent focus-loss-scheduled downgrade from firing unopposed and
-   * stranding the terminal at FOCUSED/VISIBLE/BACKGROUND mid-stream.
-   */
-  private onPtyWrite(id: string): void {
-    const managed = this.instances.get(id);
-    if (!managed) return;
-
-    managed.writeBurstDeadline = Date.now() + WRITE_BURST_DECAY_MS;
-    this.rendererPolicy.applyRendererPolicy(id, TerminalRefreshTier.BURST);
-
-    if (managed.writeBurstTimer === undefined) {
-      this.scheduleWriteBurstDecay(id, WRITE_BURST_DECAY_MS);
-    }
-  }
-
-  private scheduleWriteBurstDecay(id: string, delayMs: number): void {
-    const managed = this.instances.get(id);
-    if (!managed) return;
-    managed.writeBurstTimer = window.setTimeout(() => {
-      const current = this.instances.get(id);
-      if (!current) return;
-      current.writeBurstTimer = undefined;
-      const deadline = current.writeBurstDeadline;
-      const nowFire = Date.now();
-      if (deadline !== undefined && nowFire < deadline) {
-        this.scheduleWriteBurstDecay(id, deadline - nowFire);
-        return;
-      }
-      current.writeBurstDeadline = undefined;
-      this.rendererPolicy.applyRendererPolicy(id, current.getRefreshTier());
-    }, delayMs);
   }
 
   private onEnterPressed(id: string): void {
@@ -3503,6 +3368,7 @@ class TerminalInstanceService {
     this.cancelAttachReveal(managed);
     this.agentStateController.destroy(id);
     this.restoreController.destroy(id);
+    this.burstController.destroy(id);
 
     managed.scrollbackRestoreState = "none";
 
@@ -3544,15 +3410,6 @@ class TerminalInstanceService {
       clearTimeout(managed.tierChangeTimer);
       managed.tierChangeTimer = undefined;
     }
-    if (managed.inputBurstTimer !== undefined) {
-      clearTimeout(managed.inputBurstTimer);
-      managed.inputBurstTimer = undefined;
-    }
-    if (managed.writeBurstTimer !== undefined) {
-      clearTimeout(managed.writeBurstTimer);
-      managed.writeBurstTimer = undefined;
-    }
-    managed.writeBurstDeadline = undefined;
     if (managed.titleReportTimer !== undefined) {
       clearTimeout(managed.titleReportTimer);
       managed.titleReportTimer = undefined;
